requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.6
django-allauth==0.41.0
django-auth==0.1.9
django-cleanup==4.0.0
django-rest-framework==0.1.0
django-summernote==0.8.11.6
django-tinymce==3.0.2
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
idna==2.9
oauthlib==3.1.0
Pillow==7.1.2
psycopg2-binary==2.8.5
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
